Move IRWriter Implementation to Parser Project and Add Sulong Option to Print the IR
