+++ projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/options/SulongDebugOption.java
@@ -59,6 +59,9 @@ abstract class SulongDebugOption {
     @Option(commandLineName = "PrintLifetimeAnalysisStats", help = "Prints the results of the lifetime analysis. Can be \'true\', \'false\', \'stdout\', \'stderr\' or a filepath.", name = "printLifetimeAnalysisStatistics") //
     protected static final String PRINT_LIFE_TIME_ANALYSIS_STATS = String.valueOf(false);
 
+    @Option(commandLineName = "PrintLLVMIR", help = "Prints the LLVM IR for the parsed bitcode files, if \'true\' to stdout, if \'false\' nothing, otherwise to a file called \'<sourcefilename>.ll\'.", name = "printLLVMIR") //
+    protected static final String PRINT_LLVMIR = String.valueOf(false);
+
     @Option(commandLineName = "TraceExecution", help = "Trace execution, printing each SSA assignment. Can be \'true\', \'false\', \'stdout\', \'stderr\' or a filepath.", name = "traceExecution") //
     protected static final String TRACE_EXECUTION = String.valueOf(false);
